+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [4.1.0] - 2020-12-31
+
+### Added
+
+- Added `scroller` prop for limiting scroll distance when `mode` is set to `bottom`, in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
+
+### Fixed
+
+- Emptying container should regain stickiness, in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
+
 ## [4.0.0] - 2020-09-01
 
 ### Added